rudimentary loop point support
In Devil's Emergence, there seems to be a delay before looping, but I'm
not sure if that's a bug in this implementation or a problem with the
files I'm using. I tried using the `EndOfTrack` event to determine if
the track has ended instead of just checking if the next event index
is out of bounds, but that didn't work.